+++ packages/ui/Scroller/ScrollerBasic.js
@@ -102,6 +102,7 @@ class ScrollerBasic extends Component {
 		return x;
 	};
 
+	// scrollMode 'translate'
 	setScrollPosition (x, y) {
 		const node = this.props.scrollContentRef.current;
 
@@ -115,9 +116,9 @@ class ScrollerBasic extends Component {
 		}
 	}
 
-	// scrollMode 'translate'
-	scrollToPosition (x, y) {
-		this.props.scrollContentRef.current.scrollTo(this.getRtlPositionX(x), y);
+	// scrollMode 'native'
+	scrollToPosition (left, top, behavior) {
+		this.props.scrollContentRef.current.scrollTo({left: this.getRtlPositionX(left), top, behavior});
 	}
 
 	// scrollMode 'native'