+++ packages/ui/CHANGELOG.md
@@ -7,6 +7,14 @@ The following is a curated list of changes in the Enact ui module, newest change
 ### Removed
 - `ui/Scroller` and `ui/VirtualList` prop `data-webos-voice-focused`, `data-webos-voice-disabled`, and `data-webos-voice-group-label`
 
+### Added
+
+- `ui/MarqueeDecorator` `locale` type for `forceDirection` prop not to override the direction depending on contents
+
+## [4.1.3] - 2022-03-07
+
+- Updated to use `forwardCustom` and add `type` when forwarding custom events
+
 ## [4.1.2] - 2021-12-22
 
 ### Fixed